[Рейтинг] исправления

+++ src/main/java/club/ttg/dnd5/domain/common/repository/RatingRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface RatingRepository extends JpaRepository<Rating, String> {
     @Query("SELECT AVG(r.value) FROM Rating r WHERE r.section = :section AND r.url=:url")
-    byte getRating(String section, String url);
+    Double getRating(String section, String url);
 }
